TMDM-5255 : Mavenize 2nd pass org.talend.mdm.qizx.xmldb.
In org.talend.mdm.base, we changed only the pom.xml to download the following dependency from the online Maven repository instead of using the filesystem.

<dependency>
			    <groupId>commons-jxpath</groupId>
			    <artifactId>commons-jxpath</artifactId>
			    <version>1.3</version>
</dependency>

The reason is that org.talend.mdm.qizx.xmldb uses this dependency, and we wanted to switch it from getting it from the filesystem to getting it from the web. Because the groupId and artifactId used in base is in fact the right ones to use for downloading the dependency, we could not simply override it in org.talend.mdm.base so that the change to download it from the web is local to org.talend.mdm.qizx.xmldbb. Our change in org.talend.mdm.base might affect all projects, although theoretically it should make no difference since the same commons-jxpath jar is being downloaded as the one that was previously used in the filesystem.

git-svn-id: http://talendforge.org/svn/tom/trunk@97168 f6f1c999-d317-4740-80b0-e6d1abc6f99e